📝 Chore(#66): 코드 리뷰 반영
